Add MAAT API web client and supporting classes
This commit adds a new web client for interacting with
MAAT API, defining a method to interact with the endpoint
for getting an IoJ Appeal. The return type for this method
has not been fully implemented but this is beyond the scope
of the current work - it is simply there to act as a way
to test our connection to MAAT API.

This commit also adds supporting classes and configuration
to the service, including Resilience4j and OAuth configuration
for interacting with MAAT API. The web client itself is defined
in a new module `common`, which has been excluded from the
application module inspection as it is really a common set
of classes to be used across the modules of the service.