Split CodeQL GitHub Action
This commit splits the existing CodeQL GitHub Action into
two separate actions - one to run the CodeQL check and
another to run the Dependency Review check.

Previously, the Dependency Review check was causing the
entire action to fail, because when run against anything
other than a pull request the `base_ref/head_ref` option
was not being provided.

From the docs concerning this option: "Provide custom git
references for the git base/head when performing the
comparison check. This is only used for event types other
than pull_request and pull_request_target."

Splitting up these two checks will allow them both to report
back independently and there is also no need for the Dependency
Check to be dependant on the CodeQL check, as was the case
previously.